modified:   SoPro24Team06/Controllers/AssignmentController.cs

Closes #83

See merge request lehre/sopro/2024-sose/gruppen/team-06!75

+++ Projekt/SoPro24Team06/Controllers/AssignmentTemplateController.cs
@@ -2,6 +2,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.CodeAnalysis;
+using Microsoft.VisualBasic.CompilerServices;
+using Newtonsoft.Json;
 using SoPro24Team06.Containers;
 using SoPro24Team06.Data;
 using SoPro24Team06.Enums;
@@ -10,7 +12,6 @@ using SoPro24Team06.ViewModels;
 
 namespace SoPro24Team06.Controllers
 {
-    [Authorize]
     public class AssignmentTemplateController : Controller
     {
         private readonly UserManager<ApplicationUser> _userManager;
@@ -20,7 +21,6 @@ namespace SoPro24Team06.Controllers
         private readonly ILogger<AssignmentTemplateController> _logger;
         private readonly AssignmentTemplateContainer _assignmentTemplateContainer;
         private readonly DueTimeContainer _dueTimeContainer;
-
         public AssignmentTemplateController(
             ApplicationDbContext context,
             UserManager<ApplicationUser> userManager,
@@ -46,6 +46,16 @@ namespace SoPro24Team06.Controllers
             ViewData["contracts"] = _context.Contracts.ToList();
             CreateEditAssignmentTemplateViewModel assignmentTemplateVM =
                 new CreateEditAssignmentTemplateViewModel(processId);
+            if (processId > 0)
+            {
+                ProcessTemplate processTemplate =
+                    await _processTemplateContainer.GetProcessTemplateByIdAsync(processId);
+                ViewData["processRoles"] = processTemplate.RolesWithAccess;
+            }
+            else
+            {
+                ViewData["processRoles"] = _roleManager.Roles.ToList();
+            }
             return View("~/Views/Assignments/Create.cshtml", assignmentTemplateVM);
         }
 
@@ -131,13 +141,67 @@ namespace SoPro24Team06.Controllers
                     departmentsList,
                     contractsList,
                     assigneeType,
-                    assignedRole
+                    assignedRole,
+                    TempData["startProcessViewModel"] == null || TempData["editProcessViewModel"] == null ? 0 : (int) model.processId
                 );
 
                 _logger.LogInformation(
                     "Successfully created a new assignment template with title: {Title}",
                     model.Title
                 );
+                
+                // Author: Tamas Varadi
+                // Begin
+
+                bool redirectFromProcessController = TempData["startProcessViewModel"] != null ||
+                                                     TempData["editProcessViewModel"] != null;
+                
+                if (redirectFromProcessController)
+                {
+                    at.ForContractsList.ForEach(c => c.AssignmentsTemplates = null);
+                    at.ForDepartmentsList.ForEach(d => d.AssignmentsTemplates = null);
+                    
+                    if (TempData["startProcessViewModel"] != null)
+                    {
+                        string jsonViewModel = TempData["startProcessViewModel"] as string;
+                        StartProcessViewModel startProcessViewModel =
+                            JsonConvert.DeserializeObject<StartProcessViewModel>(jsonViewModel);
+
+                        startProcessViewModel.AssignmentTemplates.Add(at);
+
+                        jsonViewModel = JsonConvert.SerializeObject(startProcessViewModel);
+                        TempData["startProcessViewModel"] = jsonViewModel;
+
+                        return RedirectToAction("Start", "Process");
+                    }
+                    
+                    if (TempData["editProcessViewModel"] != null)
+                    {
+                        string jsonViewModel = TempData["editProcessViewModel"] as string;
+                        EditProcessViewModel editProcessViewModel =
+                            JsonConvert.DeserializeObject<EditProcessViewModel>(jsonViewModel);
+
+                        Process activeProcess = await _context.Processes.FindAsync(editProcessViewModel.Id);
+                        ApplicationUser? assignee = await _userManager.FindByIdAsync(editProcessViewModel.Supervisor.Id);
+
+                        if (at.AssigneeType == AssigneeType.WORKER_OF_REF)
+                        {
+                            assignee = await _userManager.FindByIdAsync(editProcessViewModel.WorkerOfReference.Id);
+                        }
+
+                        Assignment newAssignment = at.ToAssignment(assignee);
+                        _context.Assignments.Add(newAssignment);
+                        
+                        activeProcess.Assignments.Add(newAssignment);
+
+                        await _context.SaveChangesAsync();
+
+                        return RedirectToAction("Edit", "Process", new { id = editProcessViewModel.Id });
+
+                    }
+                }
+                
+                // End
 
                 var processTemplate = await _processTemplateContainer.GetProcessTemplateByIdAsync(
                     (int)model.processId
@@ -156,18 +220,34 @@ namespace SoPro24Team06.Controllers
 
         public async Task<IActionResult> Edit(int id)
         {
-            string queryProcessId = HttpContext.Request.Query["processId"].ToString();
-            int processId = Int32.Parse(queryProcessId);
             ViewData["roles"] = _roleManager.Roles.ToList();
             ViewData["dueIns"] = _context.DueTimes.ToList();
             ViewData["departments"] = _context.Departments.ToList();
             ViewData["contracts"] = _context.Contracts.ToList();
+            
             AssignmentTemplate assignmentTemplate =
                 _assignmentTemplateContainer.GetAssignmentTemplate(id);
+            /*
+            ProcessTemplate processTemplate= await _processTemplateContainer.GetProcessTemplateByIdAsync((int)assignmentTemplate.ProcessTemplateId);
+            ViewData["processRoles"] = processTemplate.RolesWithAccess; 
+            */
+            
+            // Überprüfen ob man das AssignmentTemplate einer ProcessTemplate oder Process bearbeitet
+            if (assignmentTemplate.ProcessTemplateId > 0)
+            {
+                ProcessTemplate processTemplate =
+                    await _processTemplateContainer.GetProcessTemplateByIdAsync((int)assignmentTemplate.ProcessTemplateId);
+                ViewData["processRoles"] = processTemplate.RolesWithAccess;
+            }
+            else
+            {
+                ViewData["processRoles"] = _roleManager.Roles.ToList();
+            }
+            
             if (assignmentTemplate != null)
             {
                 CreateEditAssignmentTemplateViewModel createEditAssignmentTemplateVM =
-                    new CreateEditAssignmentTemplateViewModel(assignmentTemplate, processId);
+                    new CreateEditAssignmentTemplateViewModel(assignmentTemplate, assignmentTemplate.ProcessTemplateId);
                 return View("~/Views/Assignments/Edit.cshtml", createEditAssignmentTemplateVM);
             }
             else
@@ -268,7 +348,32 @@ namespace SoPro24Team06.Controllers
                     "Successfully edited the assignment template with title: {Title}",
                     model.Title
                 );
-                return RedirectToAction("Detail", "ProcessTemplate", new { id = model.processId });
+                
+                // Author: Tamas Varadi
+                // Begin
+                if (model.processId == 0 || model.processId == null)
+                {
+                    AssignmentTemplate at = _assignmentTemplateContainer.GetAssignmentTemplate(model.Id);
+                    
+                    if (TempData["startProcessViewModel"] != null)
+                    {
+                        string jsonViewModel = TempData["startProcessViewModel"] as string;
+                        StartProcessViewModel startProcessViewModel =
+                            JsonConvert.DeserializeObject<StartProcessViewModel>(jsonViewModel);
+                        
+                        startProcessViewModel.AssignmentTemplates.RemoveAll(a => a.Id == model.Id);
+                        startProcessViewModel.AssignmentTemplates.Add(at);
+
+                        jsonViewModel = JsonConvert.SerializeObject(startProcessViewModel);
+                        TempData["startProcessViewModel"] = jsonViewModel;
+
+                        return RedirectToAction("Start", "Process");
+                    }
+                }
+
+                // End
+                
+                return RedirectToAction("Edit", "ProcessTemplate", new { id = model.processId });
             }
             catch (Exception ex)
             {
@@ -276,13 +381,19 @@ namespace SoPro24Team06.Controllers
                 return RedirectToAction("Edit", model);
             }
         }
-
+        [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
             AssignmentTemplate? assignmentTemplate =
                 _assignmentTemplateContainer.GetAssignmentTemplate(id);
-            _assignmentTemplateContainer.DeleteAssignmentTemplate(id);
-            return RedirectToAction(nameof(Index));
+            ProcessTemplate processTemplate = await _processTemplateContainer.GetProcessTemplateByIdAsync((int)assignmentTemplate.ProcessTemplateId);
+            foreach(var role in processTemplate.RolesWithAccess){
+                if(User.IsInRole(role.Name)|| User.IsInRole("Administrator")){
+                _assignmentTemplateContainer.DeleteAssignmentTemplate(id);
+                break;
+                }
+            }
+            return RedirectToAction("Detail", "ProcessTemplate", new { id = assignmentTemplate.ProcessTemplateId });
         }
     }
 }
